requirements.txt.txt
streamlit
requests
pandas
python-docx
python-pptx
beautifulsoup4
yt-dlp
xlsxwriter